# manufacturing-ai-cost-platform/docs/API_CONTRACT.yaml
openapi: 3.0.3
info:
  title: Manufacturing AI Cost Intelligence API
  version: 1.2.0
  description: >
    API contract for the Manufacturing AI Cost Intelligence & Autonomous
    Optimization Platform.

    Field types and persistence semantics follow DATABASE_SCHEMA.md. Cost values
    are always labelled with their provenance (ACTUAL, ESTIMATED, FORECAST,
    SIMULATED, UNAVAILABLE) so they are never presented interchangeably.
servers:
  - url: /api/v1
...
paths:
...
  /budgets:
    get:
      tags: [Budgets]
      summary: List budgets
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Budget list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Budgets]
      summary: Create a budget
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetCreate"
      responses:
        "201":
          description: Budget created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/InternalError"
  /budgets/{id}:
    patch:
      tags: [Budgets]
      summary: Update a budget
      description: >
        Budget changes are auditable events (SECURITY.md section 16).
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: Budget updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/InternalError"
  /budgets/status:
    get:
      tags: [Budgets]
      summary: Get budget status
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Budget status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetStatusList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: >
            Safe human-readable summary. Never contains stack traces, internal
            exception detail or secrets.
        request_id:
          type: string
        details:
          type: object
          description: Field-level validation detail where applicable.
    PageInfo:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    Budget:
      type: object
      properties:
        id:
          type: string
        scope_type:
          type: string
          enum: [ENTERPRISE, TENANT, PLANT, DEPARTMENT, WORKLOAD, AGENT, MODEL]
        scope_id:
          type: string
        amount:
          type: number
        currency:
          type: string
        period:
          type: string
          enum: [DAILY, MONTHLY, QUARTERLY, ANNUAL]
        warning_threshold_percent:
          type: number
        critical_threshold_percent:
          type: number
        status:
          type: string
    BudgetList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Budget"
        page:
          $ref: "#/components/schemas/PageInfo"
    BudgetCreate:
      type: object
      required:
        - scope_type
        - scope_id
        - amount
        - currency
        - period
      properties:
        scope_type:
          type: string
          enum: [ENTERPRISE, TENANT, PLANT, DEPARTMENT, WORKLOAD, AGENT, MODEL]
        scope_id:
          type: string
        amount:
          type: number
          minimum: 0
        currency:
          type: string
          default: INR
        period:
          type: string
          enum: [DAILY, MONTHLY, QUARTERLY, ANNUAL]
        warning_threshold_percent:
          type: number
          default: 80
        critical_threshold_percent:
          type: number
          default: 95
    BudgetUpdate:
      type: object
      properties:
        amount:
          type: number
          minimum: 0
        period:
          type: string
          enum: [DAILY, MONTHLY, QUARTERLY, ANNUAL]
        warning_threshold_percent:
          type: number
        critical_threshold_percent:
          type: number
        status:
          type: string
        reason:
          type: string
          description: Recorded on the audit event for this change.
    BudgetStatus:
      type: object
      properties:
        budget_id:
          type: string
        scope_type:
          type: string
        scope_id:
          type: string
        amount:
          type: number
        consumed_actual_cost:
          type: number
        consumed_estimated_cost:
          type: number
        consumed_percent:
          type: number
        currency:
          type: string
        threshold_state:
          description: >
            Budget state derived from consumption against the budget's
            thresholds. Inclusive lower bounds: consumption exactly at the
            warning threshold is WARNING. EXCEEDED is fixed at 100% and is not
            configurable.

            Null when the budget could not be compared against recorded spend
            at all - for example a currency the configured conversion policy
            cannot reach (DATABASE_SCHEMA.md section 15). Null is not a passing
            state; unevaluable_reason carries the cause, and budget evaluation
            treats the same condition as REQUIRE_APPROVAL rather than ALLOW.
          type: string
          nullable: true
          enum: [NORMAL, WARNING, CRITICAL, EXCEEDED]
        unevaluable_reason:
          type: string
          nullable: true
          description: >
            Machine-readable cause when threshold_state is null.
    BudgetStatusList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BudgetStatus"
        page:
          $ref: "#/components/schemas/PageInfo"
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authorization failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found or not visible to the caller
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"